move test_app_help_no_docs to test_plugins and refactor install_temp_app fixture
